requirements.txt
azure-mgmt-keyvault==0.30.0rc6
azure-mgmt-resource==0.30.0rc6
haikunator